moved UserInfoContainer out of 'components' folder and back into local 'Userprofile' folder
